VAC 패턴 적용(pages/home/) #9
