+++ my-diary/src/pages/home/IconArea/iconArea.style.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const DesktopCont = styled.div`
+    padding: 20px;
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+    position:absolute;
+`
